Add files via upload
The dataset in file Restaurant tips Project.xlsx contains tips data for different customers. The following are the features in the dataset:
sex -	Gender of the customer
smoker-	Indicates if the customer is a smoker or not
day-	 Day of the restaurant visit
time- Indicates whether the tip was for lunch or dinner
size	- Number of members dining
total - bill	Bill amount in USD
tip- Tip amount in USD

The following project tasks are performed in excel:
•	Used the restaurant tips file for the analytics using Excel
•	Checked if there are any missing values and clean the data
•	Identified which predictive problem is needed.
•	Encoded the categorical variables to numeric values using IF conditions
•	Built an appropriate model with the dataset. 
•	Calculated the predicted and actual tips values.
•	Calculated the RMSE(Root Mean Square Error) of the model. RMSE is root of mean of square errors.

Tools required: Microsoft Excel, Data Analysis Add-in.

Expected Deliverables were:  Model to predict restaurant tips given input values with the mathematical equation for predicting the tips value.